requirements.txt
django=>2.0.1
django-chartjs